+++ eras/shelley/impl/CHANGELOG.md
@@ -96,6 +96,8 @@
 ### `testlib`
 
 * Remove `shelleyAccountsToUMap` corresponding to the removal of `UMap` from core.
+* Added `shelleyGenRegTxCert`
+* Added `genRegTxCert` to `ShelleyEraImp`
 * Added `impSatisfySignature` and `impSatisfyMNativeScripts`
 * Added `EraSpecificSpec ShelleyEra` instance
 * Added `EraSpecificSpec` class